chore: create message package and move enums into that
