+++ src/main/java/vendingmachine/utils/message/SystemMessage.java
@@ -1,4 +1,4 @@
-package vendingmachine.utils;
+package vendingmachine.utils.message;
 
 public enum SystemMessage {
 	GET_TOTAL_AMOUNT_OF_VENDING_MACHINE("자판기가 보유하고 있는 금액을 입력해 주세요."),
